moved the case and concern requests to contracts
