+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Integration/ConcernsRecordIntegrationTests.cs
@@ -1,4 +1,5 @@
 ﻿using AutoFixture;
+using ConcernsCaseWork.API.Contracts.Case;
 using ConcernsCaseWork.API.Contracts.Concerns;
 using ConcernsCaseWork.API.Contracts.Enums;
 using ConcernsCaseWork.API.Factories;
