+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Integration/NoticeToImproveIntegrationTests.cs
@@ -1,6 +1,6 @@
 ﻿using AutoFixture;
+using ConcernsCaseWork.API.Contracts.Case;
 using ConcernsCaseWork.API.Contracts.NoticeToImprove;
-using ConcernsCaseWork.API.RequestModels;
 using ConcernsCaseWork.API.ResponseModels;
 using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.NoticeToImprove;
 using ConcernsCaseWork.API.Tests.Fixtures;
