+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/Case/CreateConcernsCase.cs
@@ -1,5 +1,5 @@
+using ConcernsCaseWork.API.Contracts.Case;
 using ConcernsCaseWork.API.Factories;
-using ConcernsCaseWork.API.RequestModels;
 using ConcernsCaseWork.API.ResponseModels;
 using ConcernsCaseWork.Data.Gateways;
